[demo] Rename markerPage to marker
